backend/requirements.txt
anthropic>=0.42.0
fastapi>=0.115.0
numpy>=2.0.0
python-dotenv>=1.0.1
uvicorn>=0.32.0
